+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `:vips` will be documented in this file.
 ### Added
 - add Image::newInterpolator() [Kleis Auke Wolthuizen]
 - implement array access interface [John Cupitt]
+- add BlendMode and Image::composite [John Cupitt]
 
 ### Deprecated
 - Nothing